+++ src/app.controller.ts
@@ -1,12 +1,19 @@
-import { Controller, Get } from '@nestjs/common';
+import { Controller, Get, Render } from '@nestjs/common';
 import { AppService } from './app.service';
+import { Public } from './global/common/decorator/skip-auth.decorator';
 
 @Controller()
 export class AppController {
   constructor(private readonly appService: AppService) {}
 
+  @Public()
   @Get()
-  getHello(): string {
-    return this.appService.getHello();
+  @Render('login.ejs')
+  async view() {
+    return { name: 'peter', age: 28, job: 'software engineer' };
   }
+
+  @Get('index')
+  @Render('order-layout.ejs')
+  async index() {}
 }
