+++ src/views/order-layout.ejs
@@ -148,8 +148,33 @@
         document.getElementById('bb').innerHTML = temp;
       })
       .catch(( error ) => {
-        // 예외처리
-        // alert(`${ error.response.data.error }`);
+        // 예외처리 - 로그인안하고 들어올때 or 로그인 쿠키가 없을 때
+        console.log(error)
+        if(error.response.status === 401){
+          alert('로그인하셔야 합니다.');
+          window.location.href = '/'
+          return
+        }
+        // 404 구매내역이 없을 때
+        else if(error.response.status === 404){
+          let temp = "";
+           temp += `
+                    <div class="container-fluid" style="border: 1px solid red; margin-top: 20px">
+                     <div class="row">
+                      <div class="col-md-3" style="border-right: 1px solid red; padding: 0">
+                        <img src="img/1296285.jpg" alt="spcFuck" style="width: 100%; margin: 0" />
+                       </div>
+                    <div class="col-md-9">
+                <h3>구매하신 상품이 없습니다. ㅎㅎㅎ</h3>
+                <h4>구매해</h4>
+                <p>바보</p>
+                <span>데데데</span>
+                <span></span>
+            </div>
+        </div>
+      </div>`;
+            document.getElementById('bb').innerHTML = temp;
+        }
       });
   }
     // 판매내역
@@ -157,7 +182,6 @@
         axios
       .get('http://localhost:3000/orders/mySellList')
       .then(( res ) => {
-        console.log("@@@@@@@@@@@@@@@@", res);
         const aa = res.data;
         let temp = "";
         for ( let i = 0; i < res.data.data.length; i++ ) {
@@ -180,9 +204,63 @@
         document.getElementById('bb').innerHTML = temp;
       })
       .catch(( error ) => {
-        // 예외처리
-        // alert(`${ error.response.data.error }`);
+        console.log(error)
+        if(error.response.status === 401){
+          alert('로그인하셔야 합니다.');
+          window.location.href = '/'
+          return
+        } else if(error.response.status === 404){
+          let temp = "";
+           temp += `
+                    <div class="container-fluid" style="border: 1px solid red; margin-top: 20px">
+                     <div class="row">
+                      <div class="col-md-3" style="border-right: 1px solid red; padding: 0">
+                        <img src="img/1296285.jpg" alt="spcFuck" style="width: 100%; margin: 0" />
+                       </div>
+                    <div class="col-md-9">
+                <h3>아직 파신 물건이 없내용</h3>
+                <h4>ㅎㅎㅎㅎㅎㅎ</h4>
+                <p>asdfasdfasf</p>
+                <span>asdfasdfasf</span>
+                <span></span>
+            </div>
+        </div>
+      </div>`;
+            document.getElementById('bb').innerHTML = temp;
+        }
+      });
+  }
+
+  function deal(){
+    axios
+      .get('http://localhost:3000/orders/myPick')
+      .then(( res ) => {
+        console.log(res.data.data.length !== 0)
+        if(res.data.data.length !== 0){
         let temp = "";
+        for ( let i = 0; i < res.data.data.length; i++ ) {
+          temp += `
+                    <div class="container-fluid" style="border: 1px solid red; margin-top: 20px">
+                     <div class="row">
+                      <div class="col-md-3" style="border-right: 1px solid red; padding: 0">
+                        <img src="img/1296285.jpg" alt="spcFuck" style="width: 100%; margin: 0" />
+                       </div>
+                    <div class="col-md-9">
+                <h3>${ res.data.data[i].product.title }</h3>
+<!--                <p>${ res.data.data[i].buyerId }</p>-->
+                <h4>${ res.data.data[i].deal }원</h4>
+                <p>날짜: ${ res.data.data[i].product.createdAt }회</p>
+                <span>조회: ${ res.data.data[i].product.viewCount }회</span>
+                <span style="float: right;">🎯 ${0} ❤ ${res.data.data[i].product.likes}</span>
+            </div>
+        </div>
+      </div>`;
+        }
+        document.getElementById('bb').innerHTML = temp;
+          }
+      })
+        .catch(( error ) => {
+          let temp = "";
            temp += `
                     <div class="container-fluid" style="border: 1px solid red; margin-top: 20px">
                      <div class="row">
@@ -199,9 +277,9 @@
         </div>
       </div>`;
             document.getElementById('bb').innerHTML = temp;
-      });
-  }
+        })
 
+}
 </script>
 
 <body>
@@ -264,8 +342,8 @@
             <div class="col" style="border: 1px solid black; height: 100px; width: 20%">
                 상품등록
             </div>
-            <div class="col" style="border: 1px solid black; height: 100px; width: 20%">
-                쌍방딜
+            <div id='deal' class="col" style="border: 1px solid black; height: 100px; width: 20%">
+                <a onclick='deal()'>딜 하러 가기</a>
             </div>
             <div class="col" style="border: 1px solid black; height: 100px; width: 20%">
                 마이페이지
